batch/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
anthropic>=0.40.0
pymupdf4llm>=0.0.10
flask>=3.0.0
